Speed up test suite by parallelizing tests
