+++ SwiftFormat.xcodeproj/xcshareddata/xcschemes/SwiftFormat (Framework).xcscheme
@@ -39,7 +39,8 @@
       </MacroExpansion>
       <Testables>
          <TestableReference
-            skipped = "NO">
+            skipped = "NO"
+            parallelizable = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "01A0EAAD1D5DB4D000A0A8E3"
